* Changed structure in /src * Added .gitignore files to ignore generated css / js files inside of /src/ * Added example directive "focusable" * Added example enum "ResponseCode" * Refactored (Renamed) FormInputTextDirective to FormInputTextComponent and added +shared folder in src/components
