+++ app/src/app.component.ts
@@ -1,12 +1,11 @@
 import { Component } from '@angular/core';
 import { Router, ROUTER_DIRECTIVES } from '@angular/router';
-
-import { AuthService } from '../../services/auth.service';
-import { IsRouteActive } from '../../pipes/is-route-active.pipe';
+import {IsRouteActive} from './pipes/is-route-active.pipe';
+import {AuthService} from './services/auth.service';
 
 @Component({
     selector: 'app',
-    templateUrl: 'src/components/app/app.component.html',
+    templateUrl: 'src/app.component.html',
     styleUrls: ['styles/components/app.css'],
     providers: [],
     directives: [ROUTER_DIRECTIVES],
