+++ app/src/bootstrap.ts
@@ -2,7 +2,7 @@ import { bootstrap } from '@angular/platform-browser-dynamic';
 import { HTTP_PROVIDERS } from '@angular/http';
 
 import { APP_ROUTER_PROVIDERS } from './routes';
-import { AppComponent } from './components/app/app.component';
+import { AppComponent } from './app.component';
 
 bootstrap(AppComponent, [
     HTTP_PROVIDERS,
