+++ app/src/components/+shared/form-input-text/form-input-text.component.ts
@@ -3,10 +3,10 @@ import { Control } from '@angular/common';
 
 @Component({
     selector: 'form-input-text',
-    templateUrl: 'src/directives/form-input-text/form-input-text.directive.html'
+    templateUrl: 'src/directives/form-input-text/form-input-text.component.html'
 })
 
-export class FormInputTextDirective {
+export class FormInputTextComponent {
     @Input() inputId: string;
     @Input() inputType: string;
     @Input() inputControl: Control;
