+++ app/src/components/+shared/index.ts
@@ -1,0 +1,1 @@
+export * from './form-input-text/form-input-text.component';
