+++ app/src/directives/focusable.directive.ts
@@ -1,0 +1,13 @@
+import {Directive, ElementRef} from '@angular/core';
+
+@Directive({
+    selector: '[focusable]'
+})
+
+export class FocusableDirective {
+    constructor(private el: ElementRef) {}
+
+    setFocus() {
+        this.el.nativeElement.focus();
+    }
+}
