+++ app/src/enums/response-code.enum.ts
@@ -1,0 +1,8 @@
+export enum ResponseCode {
+    InvalidParam,
+    InvalidCredentials,
+    NotFound,
+    SessionExpired,
+    Conflict,
+    InternalServerError
+}
